Abstract Footer.
